graphics:   - increase number of bones supported for skeletal animation (8 on psp, 12 on d3d11)   - use slerp instead of nlerp to interpolate between rotations of keyframes   - add support for lighting to default skeletal animation gpu programm   - initialise bone matrices in Entity core:   - add Slerp() to Quaternion class

+++ code/core/psp/pspfswrapper.cpp
@@ -81,7 +81,7 @@ void
 PSPFSWrapper::Write(core::FileHandle fileHandle, const void* buf, unsigned int numBytes)
 {
     int bytesWritten = sceIoWrite(fileHandle.handle, buf, numBytes);
-    CE_ASSERT(bytesWritten == numBytes, "FSWrapper::Write(): failed to write to file '%i', bytes supplied '%i', bytes written '%i'\n", fileHandle.handle, numBytes, bytesWritten);
+    CE_ASSERT((unsigned int)bytesWritten == numBytes, "FSWrapper::Write(): failed to write to file '%i', bytes supplied '%i', bytes written '%i'\n", fileHandle.handle, numBytes, bytesWritten);
 }
 
 //------------------------------------------------------------------------------
